Renames opencv_util package to swri_opencv_util. Refs #231
