+++ swri_opencv_util/mainpage.dox
@@ -2,7 +2,7 @@
 \mainpage
 \htmlinclude manifest.html
 
-\b opencv_util 
+\b swri_opencv_util 
 
 <!-- 
 Provide an overview of your package.
