+++ swri_opencv_util/src/convert.cpp
@@ -27,13 +27,13 @@
 //
 // *****************************************************************************
 
-#include <opencv_util/convert.h>
+#include <swri_opencv_util/convert.h>
 
 #include <opencv2/imgproc/imgproc.hpp>
 
 #include <ros/ros.h>
 
-namespace opencv_util
+namespace swri_opencv_util
 {
   cv::Mat ToBgra8(
       const cv::Mat& mat,
